+++ crates/wdl-engine/src/eval/v1.rs
@@ -4,8 +4,14 @@ mod expr;
 mod task;
 mod workflow;
 
+use std::fs::File;
+use std::io::BufWriter;
+use std::path::Path;
+
+use anyhow::Context;
 use anyhow::Result;
 pub use expr::*;
+use serde::Serialize;
 pub use task::*;
 pub use workflow::*;
 
@@ -13,6 +19,14 @@ use super::EvaluatedTask;
 use super::EvaluationResult;
 use crate::Outputs;
 use crate::TaskExecutionResult;
+
+/// The name of the inputs file to write for each task and workflow in the
+/// outputs directory.
+const INPUTS_FILE: &str = "inputs.json";
+
+/// The name of the outputs file to write for each task and workflow in the
+/// outputs directory.
+const OUTPUTS_FILE: &str = "outputs.json";
 
 /// Represents the kind of progress made during evaluation.
 #[derive(Debug, Clone, Copy)]
@@ -83,4 +97,13 @@ impl ProgressKind<'_> {
             | Self::WorkflowCompleted { id, .. } => id,
         }
     }
+}
+
+/// Serializes a value into a JSON file.
+fn write_json_file(path: impl AsRef<Path>, value: &impl Serialize) -> Result<()> {
+    let path = path.as_ref();
+    let file = File::create(path)
+        .with_context(|| format!("failed to create file `{path}`", path = path.display()))?;
+    serde_json::to_writer_pretty(BufWriter::new(file), value)
+        .with_context(|| format!("failed to write file `{path}`", path = path.display()))
 }